backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.2
pytest==8.3.0
